Fix issues which can not boot kernel(no correctly read nand flash) of before version,Finished mainly feature which is only to boot linux kernel of bootloader,limited is non-irq serial.

+++ JZ2440/bootloader/01_bootloaderfromscratch/include/nand.h
@@ -19,4 +19,4 @@
 /* init nand flash chip */
 void nand_init(void);
 /* read nand flash */
-void nand_read(unsigned char *buf,unsigned long start_addr,int size);
+void nand_read(unsigned char *buf,unsigned int start_addr,unsigned int size);
